Fix issues showing the completed/total information for long books by using the output of #research to replacing HERE_YOUR_STATS information.

+++ stats.h
@@ -195,8 +195,8 @@ typedef struct
 	Uint32 ranging_exp_next_lev; /*!< experience level to reach next ranging level */
 
 	Uint16 researching; /*!< flag to indicate whether a player is currently researching anything or not */
-	Uint16 research_completed; /*!< if a player is currently researching anything, this value will show how much pages are already read */
-	Uint16 research_total; /*!< if a player is currently researching anything, this value show the total amount of pages to read, until the book is completely read. */
+	Uint32 research_completed; /*!< if a player is currently researching anything, this value will show how much pages are already read */
+	Uint32 research_total; /*!< if a player is currently researching anything, this value show the total amount of pages to read, until the book is completely read. */
 } player_attribs;
 
 /*	Array for skills info required by stats bar.  Stored in an array
